Add dev/prod env separation, service layer, and app versioning
- Split db.ts into domain service modules (player, score, challenge, friend)
- Add UpdateGate component for version checking on startup (fail-open)
- Convert app.json to app.config.ts for dynamic Expo config
- Add .env.development/.env.production and EAS build profiles
- Set up Supabase CLI with config.toml and timestamped migrations
- Add app_config migration for min_version/latest_version gating
- Add root package.json with db:push:dev and db:push:prod scripts
