+++ .env.example
@@ -1,2 +1,7 @@
-VITE_SUPABASE_URL=https://your-project.supabase.co
-VITE_SUPABASE_ANON_KEY=your-anon-key-here
+# Mobile app (set in mobile-app/.env.development and .env.production)
+EXPO_PUBLIC_SUPABASE_URL=https://your-project.supabase.co
+EXPO_PUBLIC_SUPABASE_ANON_KEY=your-anon-key-here
+
+# Supabase CLI — project refs for `npm run db:push:dev` / `db:push:prod`
+SUPABASE_DEV_PROJECT_REF=your-dev-project-ref
+SUPABASE_PROD_PROJECT_REF=your-prod-project-ref
