+++ mobile-app/.env.development
@@ -1,0 +1,3 @@
+# Dev Supabase project — safe for testing, can be wiped
+EXPO_PUBLIC_SUPABASE_URL=https://imppamlsbedfzputamjf.supabase.co
+EXPO_PUBLIC_SUPABASE_ANON_KEY=sb_publishable_kcgjvzkm9zU1WKawfYvgaw__Nubj1AA
