+++ mobile-app/.env.production
@@ -1,0 +1,3 @@
+# Prod Supabase project — real user data
+EXPO_PUBLIC_SUPABASE_URL=https://bnsbcbsjkpwqeqtgtchf.supabase.co
+EXPO_PUBLIC_SUPABASE_ANON_KEY=sb_publishable_7R7hdnDejtJh0HxhG3sEVQ_Vmh-8uPR
