+++ mobile-app/app.config.ts
@@ -1,0 +1,32 @@
+import { ExpoConfig, ConfigContext } from 'expo/config';
+
+const IS_DEV = process.env.APP_ENV === 'development';
+
+export default ({ config }: ConfigContext): ExpoConfig => ({
+  ...config,
+  name: IS_DEV ? 'Emoji Games (Dev)' : 'Emoji Games',
+  slug: 'mobile-app',
+  version: '1.0.0',
+  orientation: 'portrait',
+  icon: './assets/icon.png',
+  userInterfaceStyle: 'light',
+  splash: {
+    image: './assets/splash-icon.png',
+    resizeMode: 'contain',
+    backgroundColor: '#ffffff',
+  },
+  ios: {
+    supportsTablet: true,
+  },
+  android: {
+    adaptiveIcon: {
+      backgroundColor: '#E6F4FE',
+      foregroundImage: './assets/android-icon-foreground.png',
+      backgroundImage: './assets/android-icon-background.png',
+      monochromeImage: './assets/android-icon-monochrome.png',
+    },
+  },
+  web: {
+    favicon: './assets/favicon.png',
+  },
+});
