+++ mobile-app/package.json
@@ -19,6 +19,7 @@
     "@supabase/supabase-js": "2.99.1",
     "expo": "~55.0.8",
     "expo-clipboard": "~55.0.9",
+    "expo-constants": "55.0.9",
     "expo-crypto": "~55.0.10",
     "expo-linear-gradient": "~55.0.9",
     "expo-status-bar": "~55.0.4",
