+++ mobile-app/src/services/index.ts
@@ -1,0 +1,4 @@
+export { upsertPlayer, searchPlayersByNickname } from './playerService';
+export { submitScore, fetchGlobalLeaderboard, fetchFriendsLeaderboard, type LeaderboardEntry } from './scoreService';
+export { createChallenge, fetchChallenge, linkScoreToChallenge, fetchMyChallenges, fetchChallengeLeaderboard, type ChallengeWithParticipants } from './challengeService';
+export { sendFriendRequest, acceptFriendRequest, fetchFriends, type FriendRow } from './friendService';
